Fix signed/unsigned mismatch by keeping N as size_t

+++ frameworks/ngi/nGIFramework/src/nGIProcessor.cpp
@@ -278,7 +278,7 @@ kipl::base::TImage<complex<float> ,2> DpcProcessor::ComputeHarmonicImage(kipl::b
 void DpcProcessor::ComputeResultImages()
 {
 	// Compute transmission image
-	const ptrdiff_t N=static_cast<ptrdiff_t>(openbeamDC.Size());
+	const size_t N=openbeamDC.Size();
 	complex<float> *pOpenBeamDC = openbeamDC.GetDataPtr();
 	complex<float> *pSampleDC   = sampleDC.GetDataPtr();
 
@@ -293,7 +293,7 @@ void DpcProcessor::ComputeResultImages()
 	float *pPhaseContrast=phasecontrast.GetDataPtr();
 	float *pDarkField=darkfield.GetDataPtr();
 	#pragma omp parallel for
-    for (ptrdiff_t i=0; i<N; i++)
+    for (ptrdiff_t i=0; i<static_cast<ptrdiff_t>(N); i++)
     {
 		pTransmision[i]=abs(pSampleDC[i])/abs(pOpenBeamDC[i]);
 		if (mConfig.process.bUseAmplLimits) {
